[#9]Feat: 헤더 네비게이션 바 업데이트 및 Header 컴포넌트 분리
