+++ src/lib/i18n/dict.js
@@ -250,6 +250,9 @@ Object.assign(en, {
   nav_analytics: "Analytics",
   nav_my_partners: "My Partners",
   nav_payments: "Payments",
+  nav_partner_matching: "Partner Matching",
+  nav_xrp_payment: "XRP Payment",
+  nav_my_business: "My Business",
   nav_schedule: "1:1 Meetup",
   nav_about: "About Us",
   nav_partners: "Partners",
@@ -479,6 +482,9 @@ const ko = {
   nav_analytics: "분석",
   nav_my_partners: "Biz 파트너",
   nav_payments: "XRP 무역대금 결제",
+  nav_partner_matching: "파트너 매칭",
+  nav_xrp_payment: "XRP 결제",
+  nav_my_business: "내 비즈니스",
   nav_schedule: "1:1 비즈니스 밋업",
   nav_about: "회사 소개",
   nav_partners: "파트너",
